File Transfer Web App in Golang
This is a simple web application built with Golang that allows users to easily transfer files. It includes a basic user interface where users can select and upload files to the server. The files are then saved to a designated 

folder on the server, and can be downloaded by other users. The application uses the net/http and multipart packages from the Go standard library to handle file uploads.