Add go mod init for file-transfer-app
This commit adds the go mod init file for the file-transfer-app project. This will allow us to better manage dependencies and packages for the application. 

With go mod, we can now easily add and update packages as needed without worrying about version conflicts. 

This will help us streamline our development process and make it easier to maintain the application in the future.